// types.ts
export interface Root {
  data: Data
}

export interface Data {
  threaded_conversation_with_injections_v2: ThreadedConversationWithInjectionsV2
}

export interface ThreadedConversationWithInjectionsV2 {
  instructions: Instruction[]
}

export interface Instruction {
  type: string
  entries?: Entry[]
  direction?: string
}

export interface Entry {
  entryId: string
  sortIndex: string
  content: Content
}

export interface Content {
  entryType: string
  __typename: string
  itemContent?: ItemContent
  items?: Item[]
  displayType?: string
  clientEventInfo?: ClientEventInfo2
}

export interface ItemContent {
  itemType: string
  __typename: string
  value?: string
  cursorType?: string
  tweet_results?: TweetResults
  tweetDisplayType?: string
  hasModeratedReplies?: boolean
}
// ...
export interface Item {
  entryId: string
  item: Item2
}
// ...
export interface ClientEventInfo2 {
  details: Details2
}